NP-28592 Move files
