+++ src/pages/research_profile/ResearchProfile.tsx
@@ -10,13 +10,13 @@ import orcidIcon from '../../resources/images/orcid_logo.svg';
 import { useSearchRegistrations } from '../../utils/hooks/useSearchRegistrations';
 import { PageSpinner } from '../../components/PageSpinner';
 import { useFetch } from '../../utils/hooks/useFetch';
-import { SearchResults } from '../search/SearchResults';
 import { ContributorFieldNames, SpecificContributorFieldNames } from '../../types/publicationFieldNames';
 import { ExpressionStatement } from '../../utils/searchHelpers';
 import { CristinPerson } from '../../types/user.types';
 import { filterActiveAffiliations, getFullCristinName, getOrcidUri } from '../../utils/user-helpers';
 import { UrlPathTemplate } from '../../utils/urlPaths';
 import { RootState } from '../../redux/store';
+import { RegistrationSearchResults } from '../search/result_search/RegistrationSearchResults';
 
 const textContainerSx: SxProps = {
   width: '100%',
@@ -94,7 +94,7 @@ const ResearchProfile = () => {
                   {t('common.registrations')}
                 </Typography>
                 {registrations.size > 0 ? (
-                  <SearchResults searchResult={registrations} />
+                  <RegistrationSearchResults searchResult={registrations} />
                 ) : (
                   <Typography>{t('common.no_hits')}</Typography>
                 )}
