+++ src/pages/search/result_search/RegistrationSearchResults.tsx
@@ -1,15 +1,15 @@
 import { useEffect } from 'react';
 import { Box } from '@mui/material';
-import { RegistrationList } from '../../components/RegistrationList';
-import { SearchResponse } from '../../types/common.types';
-import { Registration } from '../../types/registration.types';
-import { stringIncludesMathJax, typesetMathJax } from '../../utils/mathJaxHelpers';
+import { RegistrationList } from '../../../components/RegistrationList';
+import { SearchResponse } from '../../../types/common.types';
+import { Registration } from '../../../types/registration.types';
+import { stringIncludesMathJax, typesetMathJax } from '../../../utils/mathJaxHelpers';
 
 interface SearchResultsProps {
   searchResult: SearchResponse<Registration>;
 }
 
-export const SearchResults = ({ searchResult }: SearchResultsProps) => {
+export const RegistrationSearchResults = ({ searchResult }: SearchResultsProps) => {
   useEffect(() => {
     if (
       searchResult.hits.some(
