+++ src/utils/searchHelpers.ts
@@ -1,4 +1,4 @@
-import { registrationFilters } from '../pages/search/filters/AdvancedSearchRow';
+import { registrationFilters } from '../pages/search/result_search/filters/AdvancedSearchRow';
 
 export enum SearchParam {
   From = 'from',
